moved AListSelection.xaml
