+++ AbiCALC/AbiCALC.csproj
@@ -37,4 +37,8 @@
     <ProjectReference Include="..\lib\lib.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="customUI" />
+  </ItemGroup>
+
 </Project>
