+++ lib/interfaces/IName.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AbiCALC
+namespace lib.interfaces
 {
     public interface IName
     {
